requirements.txt
streamlit
playwright
pandas
python-dotenv
groq
openpyxl
plotly